tests: Move DeleteMessageTest class to test_message_edit.py.
